refactor : 핫한 토픽룸 조회 API
