+++ STORIX-Domain/src/main/java/com/storix/domain/domains/topicroom/application/port/LoadTopicRoomPort.java
@@ -15,7 +15,5 @@ public interface LoadTopicRoomPort {
 
     boolean existsById(Long roomId);
 
-    List<TopicRoom> loadTop5PopularRooms();
-
     List<TopicRoom> findAllActiveRooms();
 }
